Move this helper so we can reuse it
